requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.85.2
h11==0.14.0
idna==3.4
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
SQLAlchemy
alembic
psycopg2
fastapi-sqlalchemy
sqlmodel
python-dotenv
mysql-connector-python